+++ README.md
@@ -2,21 +2,22 @@
 > Static http server and base task packages.
 
 ## Installation
-1. Clone/download repo
-2. `npm install`
+1. Clone/download repo https://github.com/Reable/remote-control
+2. `git checkout dev` 
+3. `npm install`
 
 ## Usage
 **Development**
 
 `npm run start:dev`
 
-* App served @ `http://localhost:8181` with nodemon
+* App served @ `http://localhost:3000` with ts-node-dev
 
 **Production**
 
 `npm run start`
 
-* App served @ `http://localhost:8181` without nodemon
+* App served @ `http://localhost:3000`
 
 ---
 
@@ -24,7 +25,7 @@
 
 Command | Description
 --- | ---
-`npm run dev` | App served @ `http://localhost:8181` with nodemon
-`npm run start` | App served @ `http://localhost:8181` without nodemon
+`npm run start:dev` | App served @ `http://localhost:3000` with nodemon
+`npm run start` | App served @ `http://localhost:3000` without nodemon
 
 **Note**: replace `npm` with `yarn` in `package.json` if you use yarn.
